fix: deflake //rs/consensus:integration_test
`slow_checkpointing_at_upgrade_boundary` accounted for ~87% of the runtime of
`//rs/consensus:integration_test` (197s of 234s locally, the other 17 tests take
30s together), which pushed successful CI runs of that target to 421-816s
against its 900s `long` timeout. Under runner load the target crossed the
deadline and got killed: all 4 flakes of the last week are `Test timed out` with
`slow_checkpointing_at_upgrade_boundary` as the only test that never finished.

Two changes, both aimed at the runtime rather than the deadline:

* Move the test into its own `//rs/consensus:slow_checkpointing_test` target so
  it owns its timeout budget instead of sharing one with 17 other tests. This
  required moving `TestRunner` out of `tests/integration.rs` into
  `tests/framework/test_runner.rs` so both crate roots can use it.

* Don't configure chain keys for this test. Generating their transcripts,
  dealings and pre-signatures with real crypto is more than half of the test's
  runtime and has nothing to do with what it asserts. `setup_subnet` now takes
  the chain key ids, and `TestRunner::without_chain_keys()` selects none.

Measured with `--runs_per_test=3` afterwards:

| target                  | before | after |
| ----------------------- | ------ | ----- |
| `integration_test`      | 234s   |  48s  |
| `slow_checkpointing_test` |  -   | 104s  |

Both targets keep `timeout = "long"`, so no target changes which CI events run
it (`--test_timeout_filters=short,moderate` excludes `long` in merge groups
either way).
