+++ rs/consensus/BUILD.bazel
@@ -246,6 +246,79 @@ rust_test(
 )
 
 rust_test(
+    name = "slow_checkpointing_test",
+    timeout = "long",
+    srcs = glob(["tests/**"]),
+    crate_root = "tests/slow_checkpointing.rs",
+    deps = [
+        # Keep sorted.
+        ":malicious_consensus",
+        "//packages/ic-error-types",
+        "//rs/artifact_pool",
+        "//rs/bitcoin/replica_types",
+        "//rs/config",
+        "//rs/consensus/certification",
+        "//rs/consensus/chain_key",
+        "//rs/consensus/dkg",
+        "//rs/consensus/idkg:malicious_idkg",
+        "//rs/consensus/utils",
+        "//rs/crypto/prng",
+        "//rs/crypto/temp_crypto",
+        "//rs/crypto/test_utils/canister_threshold_sigs",
+        "//rs/crypto/test_utils/crypto_returning_ok",
+        "//rs/crypto/test_utils/ni-dkg",
+        "//rs/crypto/test_utils/reproducible_rng",
+        "//rs/crypto/tree_hash",
+        "//rs/crypto/utils/threshold_sig_der",
+        "//rs/https_outcalls/consensus",
+        "//rs/interfaces",
+        "//rs/interfaces/certified_stream_store",
+        "//rs/interfaces/registry",
+        "//rs/interfaces/state_manager",
+        "//rs/interfaces/state_manager/mocks",
+        "//rs/limits",
+        "//rs/monitoring/logger",
+        "//rs/monitoring/metrics",
+        "//rs/protobuf",
+        "//rs/registry/fake",
+        "//rs/registry/helpers",
+        "//rs/registry/keys",
+        "//rs/registry/proto_data_provider",
+        "//rs/registry/subnet_features",
+        "//rs/registry/subnet_type",
+        "//rs/replicated_state",
+        "//rs/sys",
+        "//rs/test_utilities",
+        "//rs/test_utilities/artifact_pool",
+        "//rs/test_utilities/consensus",
+        "//rs/test_utilities/registry",
+        "//rs/test_utilities/state",
+        "//rs/test_utilities/time",
+        "//rs/test_utilities/types",
+        "//rs/types/management_canister_types",
+        "//rs/types/types",
+        "//rs/types/types_test_utils",
+        "@crate_index//:assert_matches",
+        "@crate_index//:mockall",
+        "@crate_index//:num-traits",
+        "@crate_index//:prometheus",
+        "@crate_index//:proptest",
+        "@crate_index//:prost",
+        "@crate_index//:rand",
+        "@crate_index//:rand_chacha",
+        "@crate_index//:rayon",
+        "@crate_index//:serde_cbor",
+        "@crate_index//:slog",
+        "@crate_index//:slog-async",
+        "@crate_index//:slog-envlogger",
+        "@crate_index//:slog-term",
+        "@crate_index//:strum",
+        "@crate_index//:tempfile",
+        "@crate_index//:tokio",
+    ],
+)
+
+rust_test(
     name = "payload_test",
     srcs = glob(["tests/**"]),
     crate_root = "tests/payload.rs",
